requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
sqlalchemy
bcrypt
psycopg2
flask-wtf
flask-session